+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog JSON Schema Tools
 
 
+2014-09
+
+* refacture Reader to de-reference all pointers when initialized
+* remove :exclude_root option for object to schema_hash BREAKING change if you want nesting define your schema accordingly
+* add :links option for object to schema_hash, to include the links inline in an object
+* support items definition for array type properties - BREAKING you must change old simple property definitions
+* support oneOf definitions
+
+2014-08
+
+* add $ref resolver to merge property definitions from another file
+
 2013-10
 * allow all object properties in link href placeholders => contacts/{id}/{number}
 * add base_url option to schema hash creation. Prepends an url to all links of a rendered object